requirements.txt
selenium
webdriver-manager
apscheduler
requests
python-dotenv
pytz
